First almost-working implementation.
To come:
* Better progress reporting
* Reindex after upload
* Fix cloning
* Make sure integration tests pass
* Add a new converter for a new file type

+++ common/src/main/scala/com/overviewdocs/models/DocumentSetFile2.scala
@@ -10,6 +10,6 @@ package com.overviewdocs.models
   * it makes more sense to start at the _tail_.
   */
 case class DocumentSetFile2(
-  documentSetId: Int,
+  documentSetId: Long,
   file2Id: Long
 )
